Fix fetchProj import build error
