+++ src/app/[locale]/projects/page.tsx
@@ -6,21 +6,9 @@ import { useDispatch, useSelector } from "react-redux";
 import { useTranslations } from "next-intl";
 import { ProjectState } from "../redux/store";
 import { useState, useEffect } from "react";
-import { createClient } from "@supabase/supabase-js";
 import { initialState } from "../redux/project-slice";
 import { TrashIcon } from "@heroicons/react/24/outline";
-
-const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-);
-
-export const fetchProjects = async () => {
-  const { data, error } = await supabase.from("projects").select("*");
-  if (!error && data) {
-    return data;
-  }
-};
+import { fetchProjects } from "./fetch-proj";
 
 export default function App() {
   const dispatch = useDispatch();
